Implement varying arguments.
If the program isn't supplied any numbers, it will default to 25 min
sessions and 5 minute breaks.

When a single number is entered as argument, this is used as duration
of the sessions. The break durations is then determined as the session
duration divided by 5.

Is two numbers are supplied, the first is used as session duration, and
the second as break duration.

Is any more numbers supplied, the program will fail.